MADS: NEBULAR: Separate Macintosh presentation code
Centralize most Macintosh-specific presentation code: resource
management, graphics initialization, coordinate conversion, panel
composition, frame presentation, and native popup rendering.
