+++ engines/mads/module.mk
@@ -74,6 +74,7 @@ MODULE_OBJS := \
 	nebular/main.o \
 	nebular/main_menu.o \
 	nebular/mac_menus.o \
+	nebular/mac_nebular.o \
 	nebular/mac_resources.o \
 	nebular/menus.o \
 	nebular/popup.o \
